live_api/requirements.txt
google-genai>=1.30.0
python-dotenv>=1.0.1
pyaudio>=0.2.14
openai>=1.109.0
